half

+++ kernel.h
@@ -2,8 +2,8 @@
 #ifndef KERNEL_H
 #define KERNEL_H
 
-// #include <cuda_fp16.h>
+#include <cuda_fp16.h>
 
-extern "C" void launchMatrixAdd(float* A, float* B, float* C, int size);
-extern "C" void launchMyFA1(float* O, float* Q, float* K, float* V, float* l, float* m, int Bc, int Br,int B, int H, int N, int d);
+extern "C" void launchMatrixAdd(half* A, half* B, half* C, int size);
+extern "C" void launchMyFA1(half* O, half* Q, half* K, half* V, half* l, half* m, int Bc, int Br,int B, int H, int N, int d);
 #endif // KERNEL_H
